pyproject.toml
[project]
name = "codeinventory"
version = "0.1.0"
description = "AI-powered code inventory system"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "click>=8.1.7",
    "ollama>=0.1.7",
    "rich>=13.7.0",
    "watchdog>=3.0.0",
    "pyyaml>=6.0.1",
    "gitpython>=3.1.40",
    "requests>=2.31.0",
    "flask>=3.1.0",
    "flask-cors>=5.0.1",
]

[project.scripts]
codeinventory = "codeinventory.cli.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/codeinventory"]

[tool.hatch.build]
sources = ["src"]
